// backend/src/infrastructure/database/kysely/db.types.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface SalesInvoiceLines {
  batch_id: string | null;
  discount_percent: Generated<Numeric>;
  gst_id: string | null;
  id: Generated<string>;
  invoice_id: string;
  line_total: Generated<Numeric>;
  product_id: string;
  product_unit_id: string;
  quantity: Numeric;
  rate: Numeric;
  tax_amount: Generated<Numeric>;
}

export interface SalesInvoices {
  approved_at: Timestamp | null;
  approved_by: string | null;
  created_at: Generated<Timestamp>;
  created_by: string | null;
  customer_id: string;
  dc_id: string | null;
  discount_amount: Generated<Numeric>;
  due_date: Timestamp | null;
  financial_year_id: string;
  id: Generated<string>;
  invoice_date: Generated<Timestamp>;
  invoice_number: string;
  paid_amount: Generated<Numeric>;
  so_id: string | null;
  status: Generated<string>;
  subtotal: Generated<Numeric>;
  tax_amount: Generated<Numeric>;
  total_amount: Generated<Numeric>;
  updated_at: Generated<Timestamp>;
}